refactor: move configuration to a separate module
