+++ src/init.rs
@@ -1,6 +1,6 @@
 //! Initialize the application configuration
 
-use crate::config::create_default_config;
+use crate::config::dir::create_default_config;
 use crate::db::init::initialize_db;
 
 pub async fn init(force: bool) -> anyhow::Result<()> {
